Fix potential ODR violation in ProjectionType enum typecaster
